requirements.txt
absl-py==1.4.0
cmake==3.26.3
cycler==0.11.0
dataclasses==0.8
decorator==4.4.2
dlib==19.24.1
face-alignment==1.3.5
ffmpeg==1.4
imageio==2.15.0
imageio-ffmpeg==0.4.8
importlib-resources==5.4.0
joblib==1.1.1
kiwisolver==1.3.1
llvmlite==0.36.0
matplotlib==3.3.4
mediapipe==0.8.2
MouseInfo==0.1.3
networkx==2.5.1
numba==0.53.1
numpy==1.19.3
opencv-python==4.7.0.72
pandas==1.1.5
Pillow==8.4.0
pkg_resources==0.0.0
protobuf==3.19.6
PyAutoGUI==0.9.53
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyparsing==3.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
python-dateutil==2.8.2
python3-xlib==0.15
pytweening==1.0.4
pytz==2023.3
PyWavelets==1.1.1
PyYAML==6.0
scikit-image==0.17.2
scikit-learn==0.24.2
scipy==1.5.4
six==1.16.0
threadpoolctl==3.1.0
tifffile==2020.9.3
torch==1.10.1
torchvision==0.11.2
tqdm==4.64.1
typing_extensions==4.1.1
zipp==3.6.0
